Splitted the TeleporterCommand class up and divided it into the args[0] of the command. Also made small improvements in the code of them.
